+++ server/service/UserInput.php
@@ -1739,7 +1739,7 @@ class UserInput
                     isset($action['config'][ACTION_DELETE_SCHEDULED]) && $action['config'][ACTION_DELETE_SCHEDULED]
                 ) {
                     // if the trigger type is finished and the record id exists, check for scheduled jobs with that record and move them to status deleted
-                    $result['deleted_jobs'] = $this->delete_jobs_for_record($form_data['form_type'], $action['id'], $form_data['form_fields'][ENTRY_RECORD_ID]);
+                    $result['deleted_jobs'] = $this->delete_jobs_for_record_and_action($form_data['form_type'], $action['id'], $form_data['form_fields'][ENTRY_RECORD_ID]);
                 }
 
 
@@ -1866,6 +1866,11 @@ class UserInput
                 }
             }
 
+            if ($form_data['trigger_type'] == actionTriggerTypes_deleted) {
+                // if the trigger type is deleted and the record id exists, check for scheduled jobs with that record and move them to status deleted
+                $result['deleted_jobs'] = $this->delete_jobs_for_record($form_data['form_type'], $form_data['form_fields'][ENTRY_RECORD_ID]);
+            }
+
             $end_time = microtime(true);
             $result['time'] = array(
                 "start_date" => $start_date,
@@ -1935,7 +1940,7 @@ class UserInput
      * @param int $record_id The ID of the record.
      * @return array Array with the ids of the deleted jobs
      */
-    public function delete_jobs_for_record($form_type, $action_id, $record_id)
+    public function delete_jobs_for_record_and_action($form_type, $action_id, $record_id)
     {
         $job_status_deleted = $this->db->get_lookup_id_by_value(scheduledJobsStatus, scheduledJobsStatus_deleted);
         $job_status_queued = $this->db->get_lookup_id_by_value(scheduledJobsStatus, scheduledJobsStatus_queued);
@@ -1955,6 +1960,55 @@ class UserInput
         }
         $jobs_ids = $this->db->query_db($sql, array(
             ":action_id" => $action_id,
+            ":job_status_queued" => $job_status_queued,
+            ":record_id" => $record_id
+        ));
+        foreach ($jobs_ids as $key => $value) {
+            $this->transaction->add_transaction(
+                transactionTypes_delete,
+                transactionBy_by_system,
+                $_SESSION['id_user'],
+                $this->transaction::TABLE_SCHEDULED_JOBS,
+                $value['id'],
+                false
+            );
+            $this->db->update_by_ids(
+                "scheduledJobs",
+                array(
+                    "id_jobStatus" => $job_status_deleted
+                ),
+                array('id' => $value['id'])
+            );
+        }
+        return $jobs_ids;
+    }
+
+    /**
+     * Delete scheduled jobs associated with a record.
+     *
+     * @param int $form_type The type of form (internal or external).     
+     * @param int $record_id The ID of the record.
+     * @return array Array with the ids of the deleted jobs
+     */
+    public function delete_jobs_for_record($form_type, $record_id)
+    {
+        $job_status_deleted = $this->db->get_lookup_id_by_value(scheduledJobsStatus, scheduledJobsStatus_deleted);
+        $job_status_queued = $this->db->get_lookup_id_by_value(scheduledJobsStatus, scheduledJobsStatus_queued);
+        $sql = '';
+        if ($form_type == FORM_INTERNAL) {
+            $sql = 'SELECT id
+            FROM scheduledJobs sj
+            INNER JOIN scheduledJobs_formActions sjfa ON (sj.id = sjfa.id_scheduledJobs)
+            WHERE id_jobStatus = :job_status_queued AND sjfa.id_user_input_record = :record_id';
+        } else if (
+            $form_type == FORM_EXTERNAL
+        ) {
+            $sql = 'SELECT id
+            FROM scheduledJobs sj
+            INNER JOIN scheduledJobs_formActions sjfa ON (sj.id = sjfa.id_scheduledJobs)
+            WHERE id_jobStatus = :job_status_queued AND sjfa.id_uploadRows = :record_id';
+        }
+        $jobs_ids = $this->db->query_db($sql, array(
             ":job_status_queued" => $job_status_queued,
             ":record_id" => $record_id
         ));