+++ server/service/globals.php
@@ -188,6 +188,7 @@ define('scheduledJobsStatus_done', 'done');
 define('scheduledJobsStatus_failed', 'failed');
 define('actionTriggerTypes_started', 'started');
 define('actionTriggerTypes_finished', 'finished');
+define('actionTriggerTypes_deleted', 'deleted');
 define('timePeriod_seconds', 'seconds');
 define('timePeriod_minutes', 'minutes');
 define('timePeriod_hours', 'hours');